251031/Leontev Danil/project/project 4/requirements.txt
fastapi
uvicorn
pydantic
kafka-python
aiokafka
python-dotenv
